+++ test/support/rails_app/app/models/admin.rb
@@ -1,0 +1,7 @@
+class Admin < ActiveRecord::Base
+  devise :database_authenticatable, :confirmable, :lockable, :recoverable,
+    :registerable, :rememberable, :timeoutable, :token_authenticatable,
+    :trackable, :validatable
+
+  include Devise::Async::Model
+end